# openapi.json
{"openapi":"3.1.0","info":{"title":"IELTS Writing AI","description":"AI-powered IELTS Writing Task 2 practice and grading","version":"1.0.0"},"paths":{"/workflow_1_question/prepare_materials":{"post":{"summary":"Prepare Materials","description":"Prepare learning materials based on the generated question\n\nThis endpoint is called when the user clicks \"Start Writing\" after\ngenerating a question. It prepares additional materials like outline,\nvocabulary suggestions, and sentence suggestions.","operationId":"prepare_materials_workflow_1_question_prepare_materials_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrepareMaterialsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Prepare Materials Workflow 1 Question Prepare Materials Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/workflow_1_question":{"get":{"summary":"Get Question Generator Page","description":"Render the question generator UI page","operationId":"get_question_generator_page_workflow_1_question_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/workflow_1_question/generate":{"post":{"summary":"Generate Question","description":"Generate an IELTS Writing question based on topic and band score\n\nThis has been modified to only generate the question part first,\nand then generate the rest of the components later when the user\nclicks \"Start Writing\"","operationId":"generate_question_workflow_1_question_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopicRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Generate Question Workflow 1 Question Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/workflow_2_grading/analyze":{"post":{"tags":["grading"],"summary":"Analyze Essay","description":"Analyze and grade an IELTS Writing Task 2 essay\n\nThis endpoint accepts an essay, question, and type, then:\n1. Analyzes the essay using official IELTS criteria\n2. Provides detailed feedback and scores\n3. Highlights sentences as good, wrong, or improvable\n\nReturns:\n    Detailed feedback, scores, and highlighted essay","operationId":"analyze_essay_workflow_2_grading_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GradingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GradingResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/workflow_2_grading/":{"get":{"tags":["grading"],"summary":"Get Grading Page","description":"Render the grading UI page","operationId":"get_grading_page_workflow_2_grading__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/setting.html":{"get":{"summary":"Get Setting Page","operationId":"get_setting_page_setting_html_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/practice.html":{"get":{"summary":"Get Practice Page","operationId":"get_practice_page_practice_html_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/grading.html":{"get":{"summary":"Get Grading Page","operationId":"get_grading_page_grading_html_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Explanations":{"properties":{"green":{"items":{"$ref":"#/components/schemas/GoodSentence"},"type":"array","title":"Green","default":[]},"red":{"items":{"$ref":"#/components/schemas/WrongSentence"},"type":"array","title":"Red","default":[]},"yellow":{"items":{"$ref":"#/components/schemas/ImprovableSentence"},"type":"array","title":"Yellow","default":[]}},"type":"object","title":"Explanations","description":"Model for all sentence explanations"},"GoodSentence":{"properties":{"sentence":{"type":"string","title":"Sentence"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["sentence","reason"],"title":"GoodSentence","description":"Model for good sentences"},"GradingRequest":{"properties":{"essay":{"type":"string","title":"Essay","description":"The student's essay text"},"question":{"type":"string","title":"Question","description":"The IELTS question prompt"},"types":{"type":"string","title":"Types","description":"The type of IELTS Writing Task 2 question"}},"type":"object","required":["essay","question","types"],"title":"GradingRequest","description":"Request model for IELTS essay grading"},"GradingResponse":{"properties":{"feedback":{"type":"object","title":"Feedback","description":"Detailed feedback for each criterion"},"scores":{"allOf":[{"$ref":"#/components/schemas/ScoreDetail"}],"description":"Scores for each criterion"},"average_band":{"type":"number","title":"Average Band","description":"Overall IELTS band score (0.0-9.0)"},"essay_highlights":{"type":"string","title":"Essay Highlights","description":"HTML-formatted essay with highlighting"},"highlighted_sentences":{"allOf":[{"$ref":"#/components/schemas/HighlightCategories"}],"description":"Sentences categorized by quality"},"explanations":{"allOf":[{"$ref":"#/components/schemas/Explanations"}],"description":"Detailed explanations for highlighted sentences"}},"type":"object","required":["feedback","scores","average_band","essay_highlights","highlighted_sentences","explanations"],"title":"GradingResponse","description":"Response model for IELTS essay grading"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HighlightCategories":{"properties":{"green":{"items":{"type":"string"},"type":"array","title":"Green","default":[]},"red":{"items":{"type":"string"},"type":"array","title":"Red","default":[]},"yellow":{"items":{"type":"string"},"type":"array","title":"Yellow","default":[]}},"type":"object","title":"HighlightCategories","description":"Model for sentence highlight categories"},"ImprovableSentence":{"properties":{"sentence":{"type":"string","title":"Sentence"},"issue":{"type":"string","title":"Issue"},"improved":{"type":"string","title":"Improved"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["sentence","issue","improved","reason"],"title":"ImprovableSentence","description":"Model for improvable sentences"},"PrepareMaterialsRequest":{"properties":{"question":{"type":"string","title":"Question","description":"The generated IELTS Writing question"},"topic":{"type":"string","title":"Topic","description":"The topic for the question"},"band":{"type":"string","title":"Band","description":"The target band score"},"task_type":{"type":"string","title":"Task Type","description":"The task type (TASK1 or TASK2)","default":"TASK2"},"time_limit":{"type":"integer","title":"Time Limit","description":"Time limit in minutes","default":30}},"type":"object","required":["question","topic","band"],"title":"PrepareMaterialsRequest"},"ScoreDetail":{"properties":{"Task Response":{"type":"number","title":"Task Response"},"Coherence and Cohesion":{"type":"number","title":"Coherence And Cohesion"},"Lexical Resource":{"type":"number","title":"Lexical Resource"},"Grammatical Range and Accuracy":{"type":"number","title":"Grammatical Range And Accuracy"}},"type":"object","required":["Task Response","Coherence and Cohesion","Lexical Resource","Grammatical Range and Accuracy"],"title":"ScoreDetail","description":"Model for detailed criteria scores"},"TopicRequest":{"properties":{"topic":{"type":"string","title":"Topic"},"band":{"type":"string","title":"Band"},"task_type":{"type":"string","title":"Task Type","default":"TASK2"}},"type":"object","required":["topic","band"],"title":"TopicRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WrongSentence":{"properties":{"sentence":{"type":"string","title":"Sentence"},"error":{"type":"string","title":"Error"},"correction":{"type":"string","title":"Correction"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["sentence","error","correction","reason"],"title":"WrongSentence","description":"Model for wrong/incorrect sentences"}}}}